requirements.txt
httpx>=0.24.0
loguru>=0.7.0
python-dotenv>=1.0.0

